# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.AddStockReq:
    properties:
      author:
        description: 作者
        type: string
      category:
        description: 类别,目前有[children_story,science_knowledge,art_enlightenment]
        type: string
      name:
        description: 书本名称
        type: string
      publisher:
        description: 出版社
        type: string
      quantity_added:
        description: 添加的库存数目
        type: integer
      user_id:
        description: user_id字段不为空是则为捐献接口
        type: integer
    required:
    - author
    - category
    - name
    - publisher
    - quantity_added
    type: object
  controller.AddStockResp:
    properties:
      code:
        type: integer
      data:
        properties:
          book_id:
            description: 书本ID
            type: integer
        required:
        - book_id
        type: object
      msg:
        type: string
    required:
    - code
    - msg
    type: object
  controller.Book:
    properties:
      author:
        description: 作者
        type: string
      book_id:
        description: 书本ID
        type: integer
      category:
        description: 类别
        type: string
      created_at:
        description: 入库时间
        type: string
      name:
        description: 书本名称
        type: string
      publisher:
        description: 出版社
        type: string
      stock:
        description: 库存数量
        type: integer
      stock_status:
        description: 库存状态
        type: string
    required:
    - author
    - book_id
    - category
    - created_at
    - name
    - publisher
    - stock
    - stock_status
    type: object
  controller.BookBorrowRecord:
    properties:
      book_id:
        description: 书本ID【这个你可以理解为一类书，比如《高等数学》】
        type: integer
      copy_id:
        description: "副本ID\t【这个你可以理解为具体一本书，比如《高等数学》的第一本"
        type: integer
      return_status:
        description: 归还状态,有三种[waiting_return,returned,overdue]
        type: string
      should_return_time:
        description: 应该归还的时间
        type: string
      user_id:
        description: 用户ID
        type: integer
      user_name:
        description: 用户名
        type: string
    required:
    - book_id
    - copy_id
    - return_status
    - should_return_time
    - user_id
    - user_name
    type: object
  controller.BorrowBookReq:
    properties:
      book_id:
        description: 书本ID【这个你可以理解为一类书，比如《高等数学》】
        type: integer
      borrower_id:
        description: 借阅者ID
        type: integer
      expected_return_time:
        description: 预计归还时间,格式为"2006-01-02"
        type: string
    required:
    - book_id
    - borrower_id
    - expected_return_time
    type: object
  controller.BorrowBookResp:
    properties:
      code:
        type: integer
      data:
        properties:
          book_id:
            description: 实际借阅的书本ID
            type: integer
          copy_id:
            description: "副本ID\t【这个你可以理解为具体一本书，比如《高等数学》的第一本】"
            type: integer
        required:
        - book_id
        - copy_id
        type: object
      msg:
        type: string
    required:
    - code
    - data
    - msg
    type: object
  controller.DonateRecords:
    properties:
      book_name:
        description: 书籍名称
        type: string
      donate_num:
        description: 捐赠数目
        type: integer
      donate_time:
        description: 捐赠时间
        type: string
      phone:
        description: 电话
        type: string
      user_id:
        description: 用户ID
        type: integer
      user_name:
        description: 用户名称
        type: string
    required:
    - book_name
    - donate_num
    - donate_time
    - phone
    - user_id
    - user_name
    type: object
  controller.FuzzyQueryBookStockResp:
    properties:
      code:
        type: integer
      data:
        description: 数据
        properties:
          books:
            description: 数据
            items:
              $ref: '#/definitions/controller.Book'
            type: array
          current_page:
            description: 当前页
            type: integer
          total_num:
            description: 总数量
            type: integer
          total_page:
            description: 总页数
            type: integer
        required:
        - books
        - current_page
        - total_num
        - total_page
        type: object
      msg:
        type: string
    required:
    - code
    - data
    - msg
    type: object
  controller.GenerateCaptchaResp:
    properties:
      code:
        type: integer
      data:
        properties:
          image:
            type: string
          verification_code_id:
            type: string
        required:
        - image
        - verification_code_id
        type: object
      msg:
        type: string
    required:
    - code
    - msg
    type: object
  controller.GetDonationRankingResp:
    properties:
      code:
        type: integer
      data:
        properties:
          rankings:
            items:
              $ref: '#/definitions/controller.Rank'
            type: array
        required:
        - rankings
        type: object
    required:
    - code
    - data
    type: object
  controller.ListDonateRecordsResp:
    properties:
      code:
        type: integer
      data:
        properties:
          current_page:
            description: 当前页
            type: integer
          donate_records:
            description: 捐赠记录
            items:
              $ref: '#/definitions/controller.DonateRecords'
            type: array
          total_num:
            description: 总数量
            type: integer
          total_page:
            description: 总页数
            type: integer
        required:
        - current_page
        - donate_records
        - total_num
        - total_page
        type: object
      msg:
        type: string
    required:
    - code
    - data
    - msg
    type: object
  controller.LoginResp:
    properties:
      code:
        type: integer
      data:
        properties:
          token:
            type: string
        required:
        - token
        type: object
      msg:
        type: string
    required:
    - code
    - msg
    type: object
  controller.QueryBookBorrowRecordResp:
    properties:
      code:
        type: integer
      data:
        properties:
          borrow_records:
            items:
              $ref: '#/definitions/controller.BookBorrowRecord'
            type: array
          current_page:
            description: 当前页
            type: integer
          total_num:
            description: 总数量
            type: integer
          total_page:
            description: 总页数
            type: integer
        required:
        - borrow_records
        - current_page
        - total_num
        - total_page
        type: object
      msg:
        type: string
    required:
    - code
    - data
    - msg
    type: object
  controller.QueryStatisticsBorrowRecordsResp:
    properties:
      code:
        type: integer
      data:
        properties:
          art_enlightenment_num:
            type: integer
          children_story_num:
            type: integer
          science_knowledge_num:
            type: integer
        required:
        - art_enlightenment_num
        - children_story_num
        - science_knowledge_num
        type: object
      msg:
        type: string
    required:
    - code
    - msg
    type: object
  controller.Rank:
    properties:
      donate_num:
        description: 捐赠数目
        type: integer
      donate_times:
        description: 捐赠次数
        type: integer
      updated_at:
        description: 最近捐赠时间
        type: string
      user_id:
        description: 用户ID
        type: integer
      user_name:
        description: 用户名称
        type: string
    required:
    - donate_num
    - donate_times
    - updated_at
    - user_id
    - user_name
    type: object
  controller.SearchUserResp:
    properties:
      code:
        type: integer
      data:
        properties:
          current_page:
            type: integer
          total_page:
            type: integer
          users:
            items:
              $ref: '#/definitions/controller.User'
            type: array
        required:
        - current_page
        - total_page
        - users
        type: object
      msg:
        type: string
    required:
    - code
    - data
    - msg
    type: object
  controller.UpdateBorrowStatusReq:
    properties:
      book_id:
        description: 书本ID【这个你可以理解为一类书，比如《高等数学》】
        type: integer
      copy_id:
        description: "副本ID\t【这个你可以理解为具体一本书，比如《高等数学》的第一本】"
        type: integer
      status:
        description: 要更新的状态,取值有[waiting_return,returned,overdue]
        type: string
    required:
    - book_id
    - copy_id
    - status
    type: object
  controller.UpdateBorrowStatusResp:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    required:
    - code
    - data
    - msg
    type: object
  controller.User:
    properties:
      gender:
        description: 性别
        type: string
      id:
        description: 用户ID
        type: integer
      integral:
        description: 积分
        type: integer
      is_vip:
        description: 是否是会员
        type: boolean
      name:
        description: 姓名
        type: string
      phone:
        description: 手机号
        type: string
      status:
        description: 用户状态
        type: string
      vip_levels:
        description: 会员等级
        type: string
    required:
    - gender
    - id
    - integral
    - is_vip
    - name
    - phone
    - status
    type: object
host: localhost:8989
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for a book management system.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Book Management API
  version: "1.0"
paths:
  /api/v1/book/borrow/add:
    post:
      consumes:
      - application/json
      description: 借书接口
      parameters:
      - description: 鉴权
        in: header
        name: Authorization
        required: true
        type: string
      - description: 借书请求
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/controller.BorrowBookReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.BorrowBookResp'
      summary: 借书
      tags:
      - 借书
  /api/v1/book/borrow/query:
    get:
      consumes:
      - application/json
      description: 查询借书记录
      parameters:
      - description: 鉴权
        in: header
        name: Authorization
        required: true
        type: string
      - description: 第几页
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: page_size
        required: true
        type: integer
      - description: 借阅状态的查询条件
        in: query
        name: query_status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.QueryBookBorrowRecordResp'
      summary: 查询借书记录
      tags:
      - 借书
  /api/v1/book/borrow/query_statistics:
    get:
      consumes:
      - application/json
      description: 获取统计借阅记录
      parameters:
      - description: 鉴权
        in: header
        name: Authorization
        required: true
        type: string
      - description: 示例值 "week" "month" "year"
        in: query
        name: pattern
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.QueryStatisticsBorrowRecordsResp'
      summary: 获取统计借阅记录
      tags:
      - 借书
  /api/v1/book/borrow/update_status:
    put:
      consumes:
      - application/json
      description: 更新借阅状态
      parameters:
      - description: 鉴权
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新请求
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateBorrowStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.UpdateBorrowStatusResp'
      summary: 更新借阅状态
      tags:
      - 借书
  /api/v1/book/donate/get_ranking:
    get:
      consumes:
      - application/json
      description: 获取捐赠排名
      parameters:
      - description: 鉴权
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: top
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetDonationRankingResp'
      summary: 获取捐赠排名
      tags:
      - 捐赠
  /api/v1/book/donate/list:
    get:
      consumes:
      - application/json
      description: 列出捐赠记录
      parameters:
      - description: 鉴权
        in: header
        name: Authorization
        required: true
        type: string
      - description: 当前页
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ListDonateRecordsResp'
      summary: 列出捐赠记录
      tags:
      - 捐赠
  /api/v1/book/stock/add:
    post:
      consumes:
      - application/json
      description: 添加库存接口
      parameters:
      - description: 鉴权
        in: header
        name: Authorization
        required: true
        type: string
      - description: 增加库存请求
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/controller.AddStockReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.AddStockResp'
      summary: 添加库存
      tags:
      - 库存
  /api/v1/book/stock/fuzzy_query:
    get:
      consumes:
      - application/json
      description: 模糊查询库存信息,没有任何查询条件就是直接列出数据
      parameters:
      - description: 鉴权
        in: header
        name: Authorization
        required: true
        type: string
      - description: 作者
        in: query
        name: author
        type: string
      - description: 书本ID
        in: query
        name: book_id
        type: integer
      - description: 类别
        in: query
        name: category
        type: string
      - description: 书本名称
        in: query
        name: name
        type: string
      - description: 第几页
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.FuzzyQueryBookStockResp'
      summary: 模糊查询库存信息
      tags:
      - 库存
swagger: "2.0"